Changes to Sheet handling
Added ability to import sheets of arbitrary shapes, view them in the DrawingContextBoundZoomPreview element, use decimal values for rectangle sheet width/height, and export original un-offset sheets.

+++ DeepNestLib.Core/ISvgNestConfig.cs
@@ -48,13 +48,13 @@
 
     int SaveAsFileTypeIndex { get; set; }
 
-    int SheetHeight { get; set; }
+    double SheetHeight { get; set; }
 
     int SheetQuantity { get; set; }
 
     double SheetSpacing { get; set; }
 
-    int SheetWidth { get; set; }
+    double SheetWidth { get; set; }
 
     bool Simplify { get; set; }
